# backend/internal/handlers/post_handler.go
package handlers

import (
	"fmt"
	"grapevine/internal/database"
	"grapevine/internal/models"
	"grapevine/internal/responses"
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
)

type PostHandler struct {
	postRepo     database.PostDomain
	locationRepo database.LocationDomain
}

func NewPostHandler(postRepo database.PostDomain, locationRepo database.LocationDomain) *PostHandler {
	return &PostHandler{
		postRepo:     postRepo,
		locationRepo: locationRepo,
	}
}

type createNoteRequest struct {
	Content   string   `json:"content" binding:"required"`
	MediaURLs []string `json:"media_urls"`
}

type locationUpsertInfo struct {
	GooglePlaceID string  `json:"google_place_id" binding:"required"`
	Name          string  `json:"name" binding:"required"`
	Address       string  `json:"address" binding:"required"`
	LocationType  string  `json:"location_type"`
	Lat           float64 `json:"lat" binding:"required"`
	Lng           float64 `json:"lng" binding:"required"`
}

type createReviewRequest struct {
	Content   string   `json:"content"`
	Rating    int      `json:"rating" binding:"required,min=1,max=5"`
	MediaURLs []string `json:"media_urls"`

	LocationInfo locationUpsertInfo `json:"location" binding:"required"`
}

// CreateNoteHandler creates a standalone text/media post.
// @Summary      Create a Note
// @Description  Create a post not attached to a location.
// @Tags         posts
// @Security     BearerAuth
// @Accept       json
// @Produce      json
// @Param        request body createNoteRequest true "Note payload"
// @Success      201 {object} map[string]interface{}
// @Failure      400,401,500 {object} map[string]string
// @Router       /posts/note [post]
func (h *PostHandler) CreateNoteHandler(c *gin.Context) {

	userID, err := GetUserIDAsInt(c)
	if err != nil {
		responses.WriteError(c, http.StatusUnauthorized, "unauthorized", err.Error())
		return
	}

	var req createNoteRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		responses.WriteBadRequest(c, "invalid_request", err.Error())
		return
	}

	postID, err := h.postRepo.CreateNote(c.Request.Context(), userID, req.Content, req.MediaURLs)
	if err != nil {
		responses.WriteError(c, http.StatusInternalServerError, "creation_failed", fmt.Sprintf("Failed to create note: %s", err.Error()))
		return
	}

	c.JSON(http.StatusCreated, gin.H{"post_id": postID, "message": "Note created successfully"})
}

// CreateReviewHandler upserts the location cache and saves the review.
// @Summary      Create a Review
// @Description  Saves Google location data to cache, then creates a review attached to it.
// @Tags         posts
// @Security     BearerAuth
// @Accept       json
// @Produce      json
// @Param        request body createReviewRequest true "Review payload"
// @Success      201 {object} map[string]interface{}
// @Failure      400,401,500 {object} map[string]string
// @Router       /posts/review [post]
func (h *PostHandler) CreateReviewHandler(c *gin.Context) {
	userID, err := GetUserIDAsInt(c)
	if err != nil {
		responses.WriteError(c, http.StatusUnauthorized, "unauthorized", err.Error())
		return
	}

	var req createReviewRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		responses.WriteBadRequest(c, "invalid_request", err.Error())
		return
	}

	ctx := c.Request.Context()

	locInfo := req.LocationInfo
	locationID, err := h.locationRepo.UpsertLocation(
		ctx, locInfo.GooglePlaceID, locInfo.Name, locInfo.Address, locInfo.LocationType, locInfo.Lat, locInfo.Lng,
	)
	if err != nil {
		responses.WriteError(c, http.StatusInternalServerError, "location_error", "Failed to process location data")
		return
	}

	postID, err := h.postRepo.CreateReview(
		ctx, userID, locationID, req.Rating, req.Content, req.MediaURLs,
	)
	if err != nil {
		responses.WriteError(c, http.StatusInternalServerError, "creation_failed", "Failed to create review")
		return
	}

	c.JSON(http.StatusCreated, gin.H{"post_id": postID, "message": "Review created successfully"})
}

// GetHomeFeedHandler fetches the chronological feed of both Notes and Reviews.
// @Summary      Get Home Feed
// @Description  Returns paginated posts for the main feed.
// @Tags         posts
// @Produce      json
// @Param        limit query int false "Pagination limit" default(20)
// @Param        offset query int false "Pagination offset" default(0)
// @Success      200 {array} models.FeedItem
// @Failure      500 {object} map[string]string
// @Router       /posts [get]
func (h *PostHandler) GetHomeFeedHandler(c *gin.Context) {
	limit, _ := strconv.Atoi(c.DefaultQuery("limit", "20"))
	offset, _ := strconv.Atoi(c.DefaultQuery("offset", "0"))

	feed, err := h.postRepo.GetHomeFeed(c.Request.Context(), limit, offset)
	if err != nil {
		responses.WriteError(c, http.StatusInternalServerError, "fetch_failed", "Failed to load feed")
		return
	}

	if feed == nil {
		feed = []models.FeedItem{}
	}

	c.JSON(http.StatusOK, gin.H{"data": feed})
}

// GetPostByIDHandler fetches a single post (Note or Review) by its ID.
// @Summary      Get Post by ID
// @Description  Fetches post details including media and location info.
// @Tags         posts
// @Produce      json
// @Param        id path int true "Post ID"
// @Success      200 {object} models.FeedItem
// @Failure      400,404,500 {object} map[string]string
// @Router       /posts/{id} [get]
func (h *PostHandler) GetPostByIDHandler(c *gin.Context) {
	postIDStr := c.Param("id")
	postID, err := strconv.Atoi(postIDStr)
	if err != nil {
		responses.WriteBadRequest(c, "invalid_id", "Post ID must be a number")
		return
	}

	post, err := h.postRepo.GetPostByID(c.Request.Context(), postID)
	if err != nil {
		if err.Error() == "post not found" {
			responses.WriteError(c, http.StatusNotFound, "not_found", "Post does not exist")
			return
		}
		responses.WriteError(c, http.StatusInternalServerError, "fetch_failed", fmt.Sprintf("Failed to load post: %s", err.Error()))
		return
	}

	c.JSON(http.StatusOK, gin.H{"data": post})
}